docs(app): correct dangling references and stale wording from BM25+RRF review
- retrieval.rs: drop the no-op `as f32` cast in the BM25 idf computation
  (clippy::unnecessary_cast); formula unchanged.
- commands.rs: reattach the two-index gating-rule doc comment to
  `embed_one_page` (it was living on `sync_bm25_for_page` while two
  comments pointed readers at "embed_one_page's doc comment", which
  didn't exist); `sync_bm25_for_page` keeps its own doc describing its
  narrower role.
- commands.rs: clarify a wikify test comment that implied a mutation
  to the PRODUCTION zip (in `wikify_candidates`) would fail the test —
  it actually exercises the test's own re-implemented zip and does not
  cover the production call site.
- index_updater.rs: correct a comment claiming both indexes always
  "mutate in lockstep" — a BM25-only bootstrap update can set `changed`
  without a matching dense mutation.
- examples/retrieval_eval.rs: update the header, which still described
  a Phase-0 Gemma-3-1B-only dense harness, to describe the dense +
  fused (BM25 + RRF) measurement it now performs and the
  MEMEX_EMBED_SPEC env var.

+++ app/src-tauri/src/retrieval.rs
@@ -315,7 +315,7 @@ impl Bm25Index {
             // Document frequency is the posting count for the term — there is
             // no separate `df` map to keep in sync.
             let df = postings.len() as f32;
-            let idf = ((1.0 + (n - df + 0.5) / (df + 0.5)) as f32).ln();
+            let idf = (1.0 + (n - df + 0.5) / (df + 0.5)).ln();
             // Iteration order over one term's postings does not affect any
             // score: each doc receives exactly ONE contribution per query
             // term, so a doc's accumulation order is fixed by the query-term
